Update all core serializers with fields needed for details modal
